+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `Welcome` 组件 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `XProvider` 组件 *
+  - 🌟 [mixte] 新增 `transformKeys` 函数, 用于转换传入对象的所有键名
   - 💄 [mixte] 优化 `pick` 和 `omit` 函数类型 *
 
 ## [v3.4.0-beta.2]